phyCORE-AM335x: Strip down device tree
The phyCORE-AM335x is a SOM that can be connected to different
carrierboards like PCM-953, phyBOARD-WEGA and phyBOARD-MAIA.

It is enough for the bootloader to support the SOM specific
parts and can be so used also on different carrierboards.

Removed carrierboard specific settings like led and the
second ethernet slave.

Signed-off-by: Teresa Gámez <t.gamez@phytec.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>