boards: Use nv.linux.bootargs.console for setting the console= commandline parameter
Using a dedicated variable makes it easier to change the console= parameter
without affecting other values. Also the linux.bootargs.console variable is
already used by common/console.c. Using the same variable in the environment
avoids giving duplicate console= twice to the kernel

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Suggested-by: Teresa Gámez <t.gamez@phytec.de>

Conflicts:
	arch/arm/boards/lubbock/env/nv/linux.bootargs.console
This board was added later on. So this file does not exsist in this
branch.