ARM:phyFLEX-i.MX6: append suffix 1bank
In the current naming schema the phyFLEX-i.MX6 variants which only use
one bank for memory have the suffix "1bank". The variants which use two
banks of memory have no suffix.

The patch fixes the name of the phyFLEX-i.MX6 Solo variant with 512MiB
on one bank. So the name of the image file changes from
    barebox-phytec-pbab01s-512mb.img
    barebox-phytec-phyboard-subra-512mb.img
to
    barebox-phytec-pbab01s-512mb-1bank.img
    barebox-phytec-phyboard-subra-512mb-1bank.img

This patches touches the phyFLEX-I.MX6 Kit variant and the
phyBOARD-SUBRA which shares the same phyFLEX-I.MX6 module.

Signed-off-by: Stefan Christ <s.christ@phytec.de>