ARM:phyFLEX-i.MX6: set loadaddr to start of DDR memory.
The loadaddr in the flash-headers for the phyFLEX-i.MX6 Solo/DualLite
and Dual/Quad was set to 0x20000000 (512MiB). The start of the DDR
memory in CPU's memory map is 0x10000000 (256MiB). So the ROM loader
loads the barebox image at the memory position 256MiB and higher in the
DDR memory.

This is a problem when the module doesn't have more than 256MiB of
memory. Therefore the loadaddr is set to the start of the DDR memory.

The patch was tested on a phyFLEX-i.MX6 DualLite with 1GiB RAM on two
banks and on a phyFLEX-I.MX6 Quad with 1GiB RAM on one bank.

Signed-off-by: Stefan Christ <s.christ@phytec.de>