wdog: imx-wd: Fix imx soc reset
On i.MX6 boards the reset command doesn't work, after the commit "wdog:
imx-wd: watchdog cannot be disabled again". This commit doesn't allow
negative values, but the imx reset handler passes -1 to the
imx21_watchdog_set_timeout function for a reset in 0.5 seconds.
This patch allows -1 as timeout to set the watchdog to 0.5 seconds.

Signed-off-by: Christian Hemp <c.hemp@phytec.de>